requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
gcloud==0.17.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
jws==0.1.3
oauth2client==3.0.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-dotenv==0.19.0
python-jwt==2.0.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
uuid==1.30
wcwidth==0.2.5
wit==6.0.0
